# backend/core/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateUserDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      password_confirmation:
        type: string
      username:
        minLength: 4
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
  handler.UpdateUserDTO:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      default_city:
        type: string
      name:
        type: string
      pronouns:
        type: string
    type: object
...
  handler.UserDTO:
    properties:
      avatar_url:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  handler.UserDetailsDTO:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      default_city:
        type: string
      favorite_movies:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      pronouns:
        type: string
      username:
        type: string
    type: object
...
  handler.UserMeDetailsDTO:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      default_city:
        type: string
      email:
        type: string
      favorite_movies:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      pronouns:
        type: string
      username:
        type: string
    type: object
  handler.UserStatsDTO:
    properties:
      top_genre_name:
        type: string
      total_minutes:
        type: integer
      total_movies:
        type: integer
    type: object
...
  httputil.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8003
info:
  contact:
    email: suporte@screek.com
    name: Suporte screeK
    url: http://www.screek.com/support
  description: API para a plataforma de cinema screeK.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: screeK API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Retorna o perfil público de um usuário, incluindo seus filmes favoritos.
      parameters:
      - description: ID do Usuário (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserDetailsDTO'
        "404":
          description: Usuário não encontrado
          schema:
            type: string
      summary: Busca detalhes de um usuário pelo ID
      tags:
      - Users
  /users/{id}/stats:
    get:
      consumes:
      - application/json
      description: Retorna contadores de filmes, minutos assistidos e gênero favorito.
      parameters:
      - description: ID do Usuário (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserStatsDTO'
        "404":
          description: Estatísticas não encontradas
          schema:
            type: string
      summary: Retorna estatísticas do usuário
      tags:
      - Users
  /users/me:
    delete:
      consumes:
      - application/json
      description: Deleta permanentemente o usuário. Exige confirmação de senha.
      parameters:
      - description: JSON com a senha
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
        "401":
          description: Senha incorreta ou não autorizado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove a conta do usuário logado
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retorna os detalhes do usuário baseados no token JWT fornecido.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserMeDetailsDTO'
        "401":
          description: Não autorizado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Retorna o perfil do usuário logado
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Permite alterar Nome, Bio, AvatarURL, Pronomes e Cidade do usuário
        autenticado.
      parameters:
      - description: Dados para atualização
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
        "401":
          description: Não autorizado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Atualiza o perfil do usuário logado
      tags:
      - Users
  /users/me/favorites/{tmdb_id}:
    delete:
      description: Desvincula um filme do TMDB do perfil do usuário logado.
      parameters:
      - description: ID do filme no TMDB
        in: path
        name: tmdb_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Não autorizado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove um filme dos favoritos
      tags:
      - Users
    post:
      description: Vincula um filme do TMDB ao perfil do usuário logado.
      parameters:
      - description: ID do filme no TMDB
        in: path
        name: tmdb_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Não autorizado
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Adiciona um filme aos favoritos
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário no sistema com name, email, username e password.
      parameters:
      - description: Dados do Usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UserDTO'
        "400":
          description: Erro na validação ou usuário já existe
          schema:
            type: string
      summary: Registra um novo usuário
      tags:
      - Users
  /users/search:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de usuários que coincidem com o termo de busca
        'q'.
      parameters:
      - description: Termo de busca
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserDTO'
            type: array
        "400":
          description: Parâmetro 'q' obrigatório
          schema:
            type: string
      summary: Busca usuários por nome ou username
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"